+++ javai-persistence/src/test/java/dev/xtrafe/javai/persistence/TestKeyedCatalogueRepository.java
@@ -1,0 +1,5 @@
+package dev.xtrafe.javai.persistence;
+
+/** Repository over {@link TestKeyedCatalogue}. */
+interface TestKeyedCatalogueRepository extends JavAIRepository<TestKeyedCatalogue> {
+}